Revert "Move `matmul` and `SDPA` tests into separate files (#2806)"
This reverts commit 4fa45cecda2de19319e11e803812f2bf9ffa7365.
